Now exits with an error code if the packing failed. Also fixed a bug where it wasn't heeding the command-line arguments.

+++ scripts/support/grammar-packer.pl
@@ -21,7 +21,7 @@ my %opts = (
   m => '4g',    # amount of memory to give the packer
   c => '',       # use alternate packer config
 );
-getopts("m:c:");
+getopts("m:c:", \%opts);
 
 my $JOSHUA = $ENV{JOSHUA} or die "you must defined \$JOSHUA";
 my $CAT    = "$JOSHUA/scripts/training/scat";
@@ -48,12 +48,19 @@ print CONFIG "slice_size 100000\n\nquantizer   float   $feature_str\n";
 close(CONFIG);
 
 # Do the packing using the config.
-system("java -Xmx$opts{m} -cp $JOSHUA/class joshua.tools.GrammarPacker -c $packer_config -p $output_dir -g grammar-labeled.gz");
+my $cmd = "java -Xmx$opts{m} -cp $JOSHUA/class joshua.tools.GrammarPacker -c $packer_config -p $output_dir -g grammar-labeled.gz";
+print STDERR "Packing with $cmd\n";
+my $retval = system($cmd);
 
-# Clean up.
-unlink("grammar-labeled.gz");
-system("mv dense_map $output_dir");
-system("mv $packer_config $output_dir/packer.config");
+if ($retval == 0) {
+  # Clean up.
+  unlink("grammar-labeled.gz");
+  system("mv dense_map $output_dir");
+  system("mv $packer_config $output_dir/packer.config");
+} else {
+  print STDERR "* FATAL: Couldn't pack the grammar.\n";
+  exit 1;
+}
 
 ################################################################################
 ## SUBROUTINES #################################################################
